requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.2.14
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-extensions==3.2.3
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.2.2
pydotplus==2.0.2
PyMySQL==1.1.1
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
